Smaller delta
